Allow client to specify number of synonyms retrieved
The original DictionaryAPI file includes confusing code/comment segments attempting to explain the format of the json data being parsed and erroneously hinges personalization of output on modifying the source code. This version allows the number of synonyms returned to be specified by the user instead. The use of repeated code segments to display the user prompt menu was also eliminated